move all activities into activities folder
