+++ src/app/activities/board-game/board-game.component.spec.ts
