+++ src/app/activities/cards/cards.component.css
