+++ src/app/activities/cards/cards.component.html
