+++ src/app/activities/cards/cards.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
-import { DialogInfoComponent } from '../dialog-info/dialog-info.component';
+import { DialogInfoComponent } from '../../dialog-info/dialog-info.component';
 
-import { ActivityService } from '../activity.service';
-import { Utils } from '../utils';
-import { CardsModel } from '../models/cards.model';
+import { ActivityService } from '../../activity.service';
+import { Utils } from '../../utils';
+import { CardsModel } from '../../models/cards.model';
 
 @Component({
   selector: 'app-cards',
