+++ src/app/activities/context/context.component.css
