+++ src/app/activities/context/context.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
-import { DialogInfoComponent } from '../dialog-info/dialog-info.component';
+import { DialogInfoComponent } from '../../dialog-info/dialog-info.component';
 import { MatTableDataSource } from '@angular/material/table';
 
-import { ActivityService } from '../activity.service';
-import { Utils } from '../utils';
+import { ActivityService } from '../../activity.service';
+import { Utils } from '../../utils';
 import { interval, Subscription } from 'rxjs';
-import { ContextModel, Card, Player, ResultsThisTurn } from '../models/context.model';
+import { ContextModel, Card, Player, ResultsThisTurn } from '../../models/context.model';
 
 @Component({
   selector: 'app-context',
