+++ src/app/activities/duels/duels.component.css
