+++ src/app/activities/duels/duels.component.html
