+++ src/app/activities/duels/duels.component.spec.ts
