+++ src/app/activities/escape-room/escape-room.component.css
