+++ src/app/activities/expandables/expandables.component.css
