+++ src/app/activities/expandables/expandables.component.html
