+++ src/app/activities/expandables/expandables.component.spec.ts
