+++ src/app/activities/expandables/expandables.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
-import { DialogInfoComponent } from '../dialog-info/dialog-info.component';
+import { DialogInfoComponent } from '../../dialog-info/dialog-info.component';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 
-import { Utils } from '../utils';
-import { ActivityService } from '../activity.service';
-import { ExpandablesModel, ExpandablePair } from '../models/expandables.model';
+import { Utils } from '../../utils';
+import { ActivityService } from '../../activity.service';
+import { ExpandablesModel, ExpandablePair } from '../../models/expandables.model';
 
 @Component({
   selector: 'app-expandables',
