+++ src/app/activities/ranking/ranking.component.css
