+++ src/app/activities/ranking/ranking.component.spec.ts
