+++ src/app/activities/reword/reword.component.css
