+++ src/app/activities/story-puzzle/sentence-view/sentence-view.component.ts
