+++ src/app/activities/story-puzzle/story-puzzle.component.html
