+++ src/app/activities/timed-cards/timed-cards.component.css
