+++ src/app/app-routing.module.ts
@@ -4,23 +4,23 @@ import { AboutComponent } from './about/about.component';
 import { ChangelogComponent } from './changelog/changelog.component';
 import { HomeComponent } from './home/home.component';
 import { ContentsComponent } from './contents/contents.component';
-import { CardsComponent } from './cards/cards.component';
+import { CardsComponent } from './activities/cards/cards.component';
 import { ContactComponent } from './contact/contact.component';
 import { NewsletterSignupComponent } from './newsletter-signup/newsletter-signup.component';
 import { CheckInComponent } from './check-in/check-in.component';
-import { DuelsComponent } from './duels/duels.component';
+import { DuelsComponent } from './activities/duels/duels.component';
 import { NotFoundComponent } from './not-found/not-found.component';
-import { RewordComponent } from './reword/reword.component';
-import { ExpandablesComponent } from './expandables/expandables.component';
-import { TimedCardsComponent } from './timed-cards/timed-cards.component';
+import { RewordComponent } from './activities/reword/reword.component';
+import { ExpandablesComponent } from './activities/expandables/expandables.component';
+import { TimedCardsComponent } from './activities/timed-cards/timed-cards.component';
 import { TeachersGuideComponent } from './teachers-guide/teachers-guide.component';
 import { RoadmapComponent } from './roadmap/roadmap.component';
 import { AbbreviationsComponent } from './abbreviations/abbreviations.component';
-import { EscapeRoomComponent } from './escape-room/escape-room.component';
-import { ContextComponent } from './context/context.component';
-import { BoardGameComponent } from './board-game/board-game.component';
-import { RankingComponent } from './ranking/ranking.component';
-import { StoryPuzzleComponent } from './story-puzzle/story-puzzle.component';
+import { EscapeRoomComponent } from './activities/escape-room/escape-room.component';
+import { ContextComponent } from './activities/context/context.component';
+import { BoardGameComponent } from './activities/board-game/board-game.component';
+import { RankingComponent } from './activities/ranking/ranking.component';
+import { StoryPuzzleComponent } from './activities/story-puzzle/story-puzzle.component';
 import { DictionariesComponent } from './dictionaries/dictionaries.component';
 import { CorporaComponent } from './corpora/corpora.component';
 import { MiscComponent } from './misc/misc.component';
